ref #130 Moved all databasing tests for grid to a new file in the db directory
